fix(monolith): rename app/mcp.py to avoid shadowing mcp package
The module name `app/mcp.py` shadows the third-party `mcp` package,
causing a circular ImportError when fastmcp internally imports `mcp`.
Rename to `mcp_app.py` and update the two import sites.

Also fix invalid YAML in httpx-cf-auth semgrep rule (unquoted colons
in pattern strings).
