+++ projects/monolith/chart/Chart.yaml
@@ -1,7 +1,7 @@
 apiVersion: v2
 name: monolith
 description: Consolidated homelab web services
-version: 0.46.1
+version: 0.46.2
 type: application
 dependencies:
   - name: cf-ingress
